backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
deep-translator==1.11.4
django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
whisper
packaging==25.0
pip-tools==7.4.1
pipdeptree==2.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyproject_hooks==1.2.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
